requirements/base.txt
django==5.0.2
django-environ==0.11.2
django-cachalot==2.6.2
django-filter==24.2
django-redis==5.4.0
django-cors-headers==4.3.1
django-celery-email==3.0.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
djoser==2.2.2
argon2-cffi==23.1.0
redis==5.0.4
celery==5.4.0
flower==2.0.1
coverage==7.5.0
